Change jinyu_potion to jade_serpent_potion. Set SVN properties on the 463 mage profiles.
git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@12601 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/sc_action.cpp
@@ -273,9 +273,9 @@ void action_t::parse_spell_data( const spell_data_t& spell_data )
       base_costs[ pd -> resource() ] = floor( pd -> cost() * player -> resources.base[ pd -> resource() ] );
 
     if ( pd -> _cost_per_second > 0 )
-      costs_per_second[ pd -> resource() ] = pd -> cost_per_second();
+      costs_per_second[ pd -> resource() ] = (int) pd -> cost_per_second();
     else
-      costs_per_second[ pd -> resource() ] = floor( pd -> cost_per_second() * player -> resources.base[ pd -> resource() ] );
+      costs_per_second[ pd -> resource() ] = (int) floor( pd -> cost_per_second() * player -> resources.base[ pd -> resource() ] );
   }
 
   for ( size_t i = 1; i <= spell_data._effects -> size(); i++ )
